+++ src/theme/selectable_text.rs
@@ -1,4 +1,4 @@
-use data::{message, user::NickColor};
+use data::{config, message, theme::hex_to_color, user::NickColor};
 
 use crate::widget::selectable_text::{Catalog, Style, StyleFn};
 
@@ -46,6 +46,24 @@ pub fn accent(theme: &Theme) -> Style {
     }
 }
 
+pub fn server(
+    theme: &Theme,
+    server: Option<&message::source::Server>,
+    config: &config::buffer::ServerMessages,
+) -> Style {
+    let color = server.and_then(|server| match server.kind() {
+        message::source::server::Kind::Join => config.join.hex.as_deref().and_then(hex_to_color),
+        message::source::server::Kind::Part => config.part.hex.as_deref().and_then(hex_to_color),
+        message::source::server::Kind::Quit => config.quit.hex.as_deref().and_then(hex_to_color),
+        message::source::server::Kind::ReplyTopic => config.topic.hex.as_deref().and_then(hex_to_color),
+    }).or_else(|| text::info(theme).color);
+
+    Style {
+        color,
+        selection_color: theme.colors().accent.high_alpha,
+    }
+}
+
 pub fn nickname(theme: &Theme, nick_color: NickColor, transparent: bool) -> Style {
     let color = text::nickname(theme, nick_color, transparent).color;
 
@@ -55,10 +73,24 @@ pub fn nickname(theme: &Theme, nick_color: NickColor, transparent: bool) -> Styl
     }
 }
 
-pub fn status(theme: &Theme, status: message::source::Status) -> Style {
+pub fn status(
+    theme: &Theme,
+    status: message::source::Status,
+    config: &config::buffer::InternalMessages,
+) -> Style {
     let color = match status {
-        message::source::Status::Success => text::success(theme).color,
-        message::source::Status::Error => text::error(theme).color,
+        message::source::Status::Success => config
+            .success
+            .hex
+            .as_deref()
+            .and_then(hex_to_color)
+            .or_else(|| text::success(theme).color),
+        message::source::Status::Error => config
+            .error
+            .hex
+            .as_deref()
+            .and_then(hex_to_color)
+            .or_else(|| text::error(theme).color),
     };
 
     Style {